# Delay the security index initial bootstrap when the index is red